+++ core/src/io/worxbend/gitea4s/model/Note.scala
@@ -1,0 +1,11 @@
+package io.worxbend.gitea4s.model
+
+import zio.json.*
+
+final case class Note(
+    commit: Option[Commit] = None,
+    message: Option[String] = None
+)
+
+object Note:
+  given JsonCodec[Note] = DeriveJsonCodec.gen[Note]